+++ src/ImageResize.js
@@ -24,6 +24,10 @@ export default class ImageResize {
             moduleClasses = options.modules.slice();
         }
 
+		if (options.parchment) {
+			this.parchment = options.parchment;
+		}
+
         // Apply options to default options
         this.options = defaultsDeep({}, options, DefaultOptions);
 
@@ -55,7 +59,7 @@ export default class ImageResize {
 
         this.modules.forEach(
             (module) => {
-                module.onCreate();
+				module.onCreate(this.parchment);
             },
         );
 
@@ -187,14 +191,10 @@ export default class ImageResize {
 
     checkImage = (evt) => {
         if (this.img) {
-            if (evt.keyCode == 46 || evt.keyCode == 8) {
+            if (evt.keyCode === 46 || evt.keyCode === 8) {
                 window.Quill.find(this.img).deleteAt(0);
             }
             this.hide();
         }
     };
-}
-
-if (window.Quill) {
-    window.Quill.register('modules/imageResize', ImageResize);
 }