+++ src/modules/Resize.js
@@ -48,6 +48,8 @@ export class Resize extends BaseModule {
 
         // listen for mousedown on each box
         box.addEventListener('mousedown', this.handleMousedown, false);
+		box.addEventListener('touchstart', this.handleMousedown, false);
+
         // add drag handle to document
         this.overlay.appendChild(box);
         // keep track of drag handle
@@ -57,13 +59,22 @@ export class Resize extends BaseModule {
     handleMousedown = (evt) => {
         // note which box
         this.dragBox = evt.target;
+
         // note starting mousedown position
-        this.dragStartX = evt.clientX;
+		if (evt.touches){
+			// for mobile devices get clientX of first touch point
+			this.dragStartX = evt.touches[0].clientX;
+		} else {
+			this.dragStartX = evt.clientX;
+		}
+
         // store the width before the drag
         this.preDragWidth = this.img.width || this.img.naturalWidth;
         // set the proper cursor everywhere
         this.setCursor(this.dragBox.style.cursor);
         // listen for movement and mouseup
+		document.addEventListener('touchend', this.handleMouseup, false);
+		document.addEventListener('touchmove', this.handleDrag, false);
         document.addEventListener('mousemove', this.handleDrag, false);
         document.addEventListener('mouseup', this.handleMouseup, false);
     };
@@ -72,6 +83,8 @@ export class Resize extends BaseModule {
         // reset cursor everywhere
         this.setCursor('');
         // stop listening for movement and mouseup
+		document.addEventListener('touchend', this.handleMouseup, false);
+		document.addEventListener('touchmove', this.handleDrag, false);
         document.removeEventListener('mousemove', this.handleDrag);
         document.removeEventListener('mouseup', this.handleMouseup);
     };
@@ -82,7 +95,13 @@ export class Resize extends BaseModule {
             return;
         }
         // update image size
-        const deltaX = evt.clientX - this.dragStartX;
+		let deltaX;
+		if (evt.touches){
+			deltaX = evt.touches[0].clientX - this.dragStartX;
+		}else{
+			deltaX = evt.clientX - this.dragStartX;
+		}
+
         if (this.dragBox === this.boxes[0] || this.dragBox === this.boxes[3]) {
             // left-side resize handler; dragging right shrinks image
             this.img.width = Math.round(this.preDragWidth - deltaX);