+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## v4.0.0 - August 12, 2023
+
+- Remove Quill register to window
+- Parchment now has to be supplied via the module configuration
+- Fixed for mobile devices
+- Dependencies updated
+
 ## v3.0.0 - April 9, 2017
 
 - Position drag handles relative to editor, not document